requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.5.0
numpy>=1.24.0
opencv-python>=4.7.0
facenet-pytorch>=2.5.3
mtcnn>=0.1.1
scipy>=1.10.0
scikit-image>=0.20.0
lpips>=0.1.4
flask>=2.3.0
flask-cors>=4.0.0
gradio>=3.50.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0
